env_requirements.txt
awesome-slugify==1.6.5
bottle==0.12.9
Pillow==3.3.1
pyvirtualdisplay==0.2
selenium==3.0.0b2
